JWT done

+++ blog/blog_post/views.py
@@ -11,6 +11,8 @@ from rest_framework.permissions import IsAdminUser, IsAuthenticated # noqa
 
 class TagCreateView(APIView):
 
+    permission_classes = [IsAuthenticated]
+
     def get(self, request):
         tags = Tag.objects.all()
         serializer = TagSerializer(tags, many=True)
@@ -19,7 +21,7 @@ class TagCreateView(APIView):
 
 class TagDetailView(APIView):
 
-    # permission_classes = [IsAdminUser]
+    permission_classes = [IsAuthenticated]
 
     def post(self, request):
         serializer = TagSerializer(data=request.data)
@@ -29,6 +31,9 @@ class TagDetailView(APIView):
         return Response(status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        self.permission_classes = [IsAdminUser]
+        self.check_permission(request)
+
         try:
             tag = Tag.objects.get(id=pk)
             tag.delete()
@@ -49,6 +54,8 @@ class CategoryListView(APIView):
 
 
 class CategoryDetailView(APIView):
+
+    permission_classes = [IsAdminUser]
 
     def post(self, request):
         serializer = CategorySerializer(data=request.data)
